Refactor `get_selected_commit` to return a specific error.
When the user has started a new repo, but there aren't any commits yet,
calling `get_selected_commit` produces a specific error message. This
change alters the function to return an application specific error (so a
good message can be displayed to the user) and changes the `From` trait
implementation for the application error class to map the `git2` class
and code values to the correct application error.